Refactor: Merge errors crate into expanded tpm2::errors module
