+++ Cargo.toml
@@ -7,7 +7,6 @@ members = [
     "crates/tpm2",
     "crates/tpm2-client",
     "crates/tpm2-impl",
-    "errors",
     "marshalable",
     "marshalable-derive",
 ]
@@ -41,7 +40,6 @@ zerocopy = { version = "0.8.33", features = ["derive"] }
 
 # Common workspace crates
 tpm2-rs-base = { path = "base" }
-tpm2-rs-errors = { path = "errors" }
 tpm2-rs-marshalable = { path = "marshalable" }
 tpm2-rs-marshalable-derive = { path = "marshalable-derive" }
 tpm2 = { path = "crates/tpm2" }
