+++ base/Cargo.toml
@@ -14,9 +14,9 @@ categories.workspace = true
 bitflags = { workspace = true }
 open-enum = { workspace = true }
 safe-discriminant = { workspace = true }
-tpm2-rs-errors = { workspace = true }
+tpm2 = { workspace = true }
 tpm2-rs-marshalable = { workspace = true }
 
 [features]
 # Implement conversions/traits which require std
-std = ["tpm2-rs-errors/std"]
+std = ["tpm2/std"]
