+++ crates/tpm2/src/errors/mod.rs
@@ -1,11 +1,12 @@
-#![forbid(unsafe_code)]
-#![no_std]
+//! Errors used throughout this base crate.
 
+use core::convert::Infallible;
 use core::convert::TryFrom;
 use core::error::Error;
 use core::fmt;
 use core::num::NonZeroU32;
 use core::result::{Result, Result::*};
+
 pub use tpm_rc::*;
 pub use tss_rc::*;
 
@@ -14,10 +15,34 @@ mod std;
 mod tpm_rc;
 mod tss_rc;
 
+/// Any error which can happen when marshalling
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct MarshalError;
+
+/// Any error which can happen when unmarshalling
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct UnmarshalError;
+
+impl From<Infallible> for UnmarshalError {
+    fn from(value: Infallible) -> Self {
+        match value {}
+    }
+}
+
+/// Specific error type corresponding to TPM_RC_HASH
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct HashError;
+
+impl From<HashError> for UnmarshalError {
+    fn from(_: HashError) -> Self {
+        Self
+    }
+}
+
 /// Represents success or [`TssError`] failure, which can happen at any layer.
 pub type TssResult<T> = Result<T, TssError>;
 
-/// A TSS error that can occur any any layer, e.g. Service (`TpmRcError`) and Client errors
+/// A TSS error that can occur at any layer, e.g. Service (`TpmRcError`) and Client errors
 /// can be coalesced into this error type.
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub struct TssError(NonZeroU32);
