+++ crates/tpm2/src/errors/std.rs
@@ -1,7 +1,7 @@
 //! Implement std-only traits and conversions
 extern crate std;
 
-use crate::TssError;
+use crate::errors::TssError;
 
 impl From<TssError> for std::io::Error {
     fn from(value: TssError) -> Self {
