--- errors/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "tpm2-rs-errors"
-# Inherit from Workspace Settings
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-homepage.workspace = true
-repository.workspace = true
-license.workspace = true
-keywords.workspace = true
-categories.workspace = true
-
-[features]
-# Implement conversions/traits which require std
-std = []
-
-[dependencies]
